# server/api/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Visora AI P1", "version": "0.2.0", "description": "本地开发账号服务与运营后台。写请求必须通过浏览器 Origin 精确白名单验证，缺失或不匹配返回403 FORBIDDEN。写方法包含 POST/PATCH/PUT/DELETE，四者都要求 Origin、X-Visora-Request 与 Content-Type 三道校验。浏览器只调用本 origin 的 /api；开发用户端/Admin统一使用127.0.0.1，不混用localhost。代理保留 Origin、Cookie 和 Set-Cookie，不让浏览器直接请求API_TARGET。项目云存储、平台生成任务与积分消耗结算仍未接入；积分流水中的 consume 类型由后续生成迭代写入，当前没有写入方。" },
  "security": [{ "session": [] }],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login", "security": [], "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/Credentials" },
        "responses": { "200": { "$ref": "#/components/responses/Login" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/auth/me": {
      "get": { "operationId": "currentUser", "description": "返回当前账号，含本人积分余额。", "responses": { "200": { "$ref": "#/components/responses/User" }, "401": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout", "security": [], "description": "Cookie可选；不存在/过期/禁用/已删除会话均可重复退出。成功总是清理同名同Path Cookie；仍需验证Origin与请求头。", "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false } } } },
        "responses": { "204": { "description": "会话已退出", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" }, "Set-Cookie": { "description": "visora_session 清理：Path=/; HttpOnly; SameSite=Strict; Max-Age=0，Secure与登录模式一致。", "schema": { "type": "string" } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/membership/applications": {
      "post": {
        "operationId": "submitMembershipApplication", "description": "已登录用户提交会员申请，进入 pending 供管理员审批。提交不做去重：同一用户可有多条待审申请，审批时逐条处理。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ApplicationSubmit" },
        "responses": { "201": { "$ref": "#/components/responses/Application" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "listUsers", "description": "仅管理员；默认及精确邮箱筛选均包含active/disabled。createdAt和_id倒序，cursor与筛选条件绑定；前端维护cursor栈提供上一页，更换查询或刷新重置cursor栈。不返回密码、sessionVersion或会话。",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "name": "email", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/UserPage" },
          "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "operationId": "createUser", "description": "仅管理员，创建普通 active 用户；拒绝客户端 role/status 字段。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/Credentials" },
        "responses": { "201": { "$ref": "#/components/responses/User" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/users/{id}": {
      "patch": {
        "operationId": "setUserStatus", "description": "仅管理员可停用/恢复普通用户；管理员账号受保护。恢复不能使旧会话复活。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/StatusChange" },
        "responses": { "200": { "$ref": "#/components/responses/User" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/points/ledger": {
      "get": {
        "operationId": "listPointEntries", "description": "仅管理员；积分流水，createdAt/_id 倒序。kind 为 grant/revoke/topup/allocate/allocation_revoke/consume。团队池流水没有 userId，只有 teamId/teamName；成员分配流水同时有 userId 与 teamId。",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "name": "userId", "in": "query", "schema": { "type": "string" } },
          { "name": "kind", "in": "query", "schema": { "type": "string", "enum": ["grant", "revoke", "topup", "allocate", "allocation_revoke", "consume"] } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/PointEntryPage" },
          "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/admin/points/grant": {
      "post": {
        "operationId": "grantPoints", "description": "仅管理员；给用户发放积分。余额与流水在同一事务内写入。单次上限 100000000，超过返回 400。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/PointMove" },
        "responses": { "200": { "$ref": "#/components/responses/PointEntry" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/points/revoke": {
      "post": {
        "operationId": "revokePoints", "description": "仅管理员；撤销用户积分。余额不足时返回 409 INSUFFICIENT_POINTS，且不会把余额截断到 0，也不会写入流水。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/PointMove" },
        "responses": { "200": { "$ref": "#/components/responses/PointEntry" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/models": {
      "get": {
        "operationId": "listModels", "description": "仅管理员；平台模型目录，createdAt/_id 倒序。目录只放报价与启停，不含供应商密钥。",
        "parameters": [{ "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "$ref": "#/components/responses/ModelPage" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "operationId": "createModel", "description": "仅管理员；名称唯一，重复返回 409 NAME_EXISTS。kind 仅接受 image/video/text/audio。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ModelCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Model" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/models/{id}": {
      "patch": {
        "operationId": "updateModel", "description": "仅管理员；局部更新，至少要带一个字段，否则 400。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ModelUpdate" },
        "responses": { "200": { "$ref": "#/components/responses/Model" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/teams": {
      "get": {
        "operationId": "listTeams", "description": "仅管理员；团队列表，createdAt/_id 倒序。allocatedThisMonth 由积分流水聚合得到（UTC 自然月）。",
        "parameters": [{ "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "$ref": "#/components/responses/TeamPage" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "operationId": "createTeam", "description": "仅管理员；创建团队并把指定用户设为团队管理员成员。团队名唯一，重复返回 409 NAME_EXISTS。新团队积分池为 0。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/TeamCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Team" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/api/admin/membership/applications": {
      "get": {
        "operationId": "listApplications", "description": "仅管理员；会员申请列表，createdAt/_id 倒序，可按 status 过滤。",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "approved", "rejected"] } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/ApplicationPage" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "headers": { "NoStore": { "schema": { "type": "string", "const": "no-store" } } },
    "securitySchemes": { "session": { "type": "apiKey", "in": "cookie", "name": "visora_session" } },
    "parameters": {
      "RequestHeader": { "name": "X-Visora-Request", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } },
...
    },
    "requestBodies": {
      "Credentials": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password"], "additionalProperties": false, "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string", "writeOnly": true } } } } } },
...
    },
...
    "responses": {
...
      "Login": { "description": "身份确认成功，设置服务端会话Cookie", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" }, "Set-Cookie": { "description": "visora_session=<opaque>; Path=/; HttpOnly; SameSite=Strict; Max-Age由显式SESSION_TTL产生。HTTPS模式含Secure，loopback HTTP开发模式不含Secure。", "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
...
      "Error": { "description": "受控错误；不回显数据库错误或凭据", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}